# backend/src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"PropertyResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"ownerId": {
						"type": "number",
						"format": "double"
					},
					"managerId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"addressId": {
						"type": "number",
						"format": "double"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"genderType": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"ownerId",
					"addressId",
					"latitude",
					"longitude",
					"genderType",
					"status",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatePropertyRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"ownerId": {
						"type": "number",
						"format": "double"
					},
					"managerId": {
						"type": "number",
						"format": "double"
					},
					"addressId": {
						"type": "number",
						"format": "double"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"genderType": {
						"type": "string",
						"enum": [
							"male",
							"female",
							"mixed"
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						]
					}
				},
				"required": [
					"name",
					"ownerId",
					"addressId",
					"latitude",
					"longitude",
					"genderType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdatePropertyRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"managerId": {
						"type": "number",
						"format": "double"
					},
					"addressId": {
						"type": "number",
						"format": "double"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"genderType": {
						"type": "string",
						"enum": [
							"male",
							"female",
							"mixed"
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						]
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"RoomResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"propertyId": {
						"type": "number",
						"format": "double"
					},
					"roomNumber": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"propertyId",
					"roomNumber",
					"price",
					"status",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateRoomRequest": {
				"properties": {
					"propertyId": {
						"type": "number",
						"format": "double"
					},
					"roomNumber": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string",
						"enum": [
							"available",
							"occupied",
							"maintenance"
						]
					}
				},
				"required": [
					"propertyId",
					"roomNumber",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateRoomRequest": {
				"properties": {
					"roomNumber": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string",
						"enum": [
							"available",
							"occupied",
							"maintenance"
						]
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"FacilityResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateFacilityRequest": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddPropertyFacilityRequest": {
				"properties": {
					"propertyId": {
						"type": "number",
						"format": "double"
					},
					"facilityId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"propertyId",
					"facilityId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "ngekost-aja-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/properties": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PropertyResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Browse / Search properties",
				"tags": [
					"Property"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "location",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "genderType",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"male",
								"female",
								"mixed"
							]
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"active",
								"inactive"
							]
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PropertyResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"description": "Create new property",
				"tags": [
					"Property"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePropertyRequest"
							}
						}
					}
				}
			}
		},
		"/properties/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PropertyResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Get property details",
				"tags": [
					"Property"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PropertyResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Update property",
				"tags": [
					"Property"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePropertyRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Delete property",
				"tags": [
					"Property"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/rooms": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RoomResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "List rooms for a property",
				"tags": [
					"Room"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "propertyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoomResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"description": "Create new room",
				"tags": [
					"Room"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateRoomRequest"
							}
						}
					}
				}
			}
		},
		"/rooms/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoomResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Get room by ID",
				"tags": [
					"Room"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoomResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Update room",
				"tags": [
					"Room"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateRoomRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Delete room",
				"tags": [
					"Room"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/facilities": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/FacilityResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "List all facilities",
				"tags": [
					"Facility"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FacilityResponse"
								}
							}
						}
					}
				},
				"description": "Create new facility",
				"tags": [
					"Facility"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateFacilityRequest"
							}
						}
					}
				}
			}
		},
		"/facilities/property": {
			"post": {
				"operationId": "AddToProperty",
				"responses": {
					"201": {
						"description": "Added",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Add facility to property",
				"tags": [
					"Facility"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddPropertyFacilityRequest"
							}
						}
					}
				}
			}
		},
		"/facilities/property/{propertyId}/{facilityId}": {
			"delete": {
				"operationId": "RemoveFromProperty",
				"responses": {
					"200": {
						"description": "Removed",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Remove facility from property",
				"tags": [
					"Facility"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "facilityId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/facilities/property/{propertyId}": {
			"get": {
				"operationId": "GetPropertyFacilities",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/FacilityResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get facilities for a property",
				"tags": [
					"Facility"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}